Release v1.3.0

+++ Projects/NUCLEO-U575ZI-Q/Applications/FreeRTOS_Mutex/Src/app_freertos.c
@@ -121,31 +121,32 @@ void MX_FREERTOS_Init(void) {
   /* USER CODE END RTOS_EVENTS */
 
 }
+
 /* USER CODE BEGIN Header_Thread1_Entry */
 /**
-* @brief Function implementing the Thread1 thread.
-* @param argument: Not used
-* @retval None
-*/
+  * @brief Function implementing the Thread1 thread
+  * @param argument: Not used
+  * @retval None
+  */
 /* USER CODE END Header_Thread1_Entry */
 void Thread1_Entry(void *argument)
 {
   /* USER CODE BEGIN Thread1 */
-  uint16_t i;
+  uint8_t i;
   /* Infinite loop */
   for(i = 0; i < 10; ++i)
   {
-    #if EXAMPLE_USES_MUTEX
-      osMutexAcquire(MutexHandle, osWaitForever);
-      printf ("Thread1: Mutex Acquired!\n");
-    #endif
+#if EXAMPLE_USES_MUTEX
+    osMutexAcquire(MutexHandle, osWaitForever);
+    printf ("Thread1: Mutex Acquired!\n");
+#endif
 
     printf("Thread1 : This is message number %u\n", i+1);
 
-    #if EXAMPLE_USES_MUTEX
-      printf ("Thread1: Mutex Released!\n");
-      osMutexRelease(MutexHandle);
-    #endif
+#if EXAMPLE_USES_MUTEX
+    printf ("Thread1: Mutex Released!\n");
+    osMutexRelease(MutexHandle);
+#endif
 
     HAL_GPIO_TogglePin(LED_GREEN_GPIO_Port, LED_GREEN_Pin);
     osDelay(200);
@@ -161,15 +162,15 @@ void Thread1_Entry(void *argument)
 
 /* USER CODE BEGIN Header_Thread2_Entry */
 /**
-* @brief Function implementing the Thread2 thread.
-* @param argument: Not used
-* @retval None
-*/
+  * @brief Function implementing the Thread2 thread
+  * @param argument: Not used
+  * @retval None
+  */
 /* USER CODE END Header_Thread2_Entry */
 void Thread2_Entry(void *argument)
 {
   /* USER CODE BEGIN Thread2 */
-  uint16_t i;
+  uint8_t i;
 
   /* Infinite loop */
   for(i = 0; i < 10; ++i)
@@ -185,6 +186,7 @@ void Thread2_Entry(void *argument)
     printf ("Thread2: Mutex Released!\n");
     osMutexRelease(MutexHandle);
 #endif
+
     HAL_GPIO_TogglePin(LED_BLUE_GPIO_Port, LED_BLUE_Pin);
     osDelay(200);
   }
